General refactoring and cleanup.
