+++ SpidNetSdk/SpidNetSdk/Saml2/Response.cs
@@ -1,0 +1,99 @@
+﻿using System;
+using System.Security.Cryptography.Xml;
+using System.Xml;
+
+namespace SpidNetSdk.Saml2
+{
+    public class Response
+    {
+        private XmlDocument xmlDoc;
+        private SamlAccountSettings accountSettings;
+        private Certificate certificate;
+
+        public Response(SamlAccountSettings accountSettings)
+        {
+            this.accountSettings = accountSettings;
+            certificate = new Certificate();
+            certificate.LoadCertificate(accountSettings.Certificate);
+        }
+
+        public void LoadXml(string xml)
+        {
+            xmlDoc = new XmlDocument();
+            xmlDoc.PreserveWhitespace = true;
+            xmlDoc.XmlResolver = null;
+            xmlDoc.LoadXml(xml);
+        }
+
+        public void LoadXmlFromBase64(string response)
+        {
+            System.Text.ASCIIEncoding enc = new System.Text.ASCIIEncoding();
+            LoadXml(enc.GetString(Convert.FromBase64String(response)));
+        }
+
+        public bool IsValid()
+        {
+            bool status = true;
+
+            XmlNamespaceManager manager = new XmlNamespaceManager(xmlDoc.NameTable);
+            manager.AddNamespace("ds", SignedXml.XmlDsigNamespaceUrl);
+            manager.AddNamespace("saml", "urn:oasis:names:tc:SAML:2.0:assertion");
+            manager.AddNamespace("samlp", "urn:oasis:names:tc:SAML:2.0:protocol");
+            XmlNodeList nodeList = xmlDoc.SelectNodes("//ds:Signature", manager);
+
+            SignedXml signedXml = new SignedXml(xmlDoc);
+            signedXml.LoadXml((XmlElement)nodeList[0]);
+
+            status &= signedXml.CheckSignature(certificate.cert, true);
+
+            var notBefore = NotBefore();
+            status &= !notBefore.HasValue || (notBefore <= DateTime.Now);
+
+            var notOnOrAfter = NotOnOrAfter();
+            status &= !notOnOrAfter.HasValue || (notOnOrAfter > DateTime.Now);
+
+            return status;
+        }
+
+        public DateTime? NotBefore()
+        {
+            XmlNamespaceManager manager = new XmlNamespaceManager(xmlDoc.NameTable);
+            manager.AddNamespace("saml", "urn:oasis:names:tc:SAML:2.0:assertion");
+            manager.AddNamespace("samlp", "urn:oasis:names:tc:SAML:2.0:protocol");
+
+            var nodes = xmlDoc.SelectNodes("/samlp:Response/saml:Assertion/saml:Conditions", manager);
+            string value = null;
+            if (nodes != null && nodes.Count > 0 && nodes[0] != null && nodes[0].Attributes != null && nodes[0].Attributes["NotBefore"] != null)
+            {
+                value = nodes[0].Attributes["NotBefore"].Value;
+            }
+            return value != null ? DateTime.Parse(value) : (DateTime?)null;
+        }
+
+        public DateTime? NotOnOrAfter()
+        {
+            XmlNamespaceManager manager = new XmlNamespaceManager(xmlDoc.NameTable);
+            manager.AddNamespace("saml", "urn:oasis:names:tc:SAML:2.0:assertion");
+            manager.AddNamespace("samlp", "urn:oasis:names:tc:SAML:2.0:protocol");
+
+            var nodes = xmlDoc.SelectNodes("/samlp:Response/saml:Assertion/saml:Conditions", manager);
+            string value = null;
+            if (nodes != null && nodes.Count > 0 && nodes[0] != null && nodes[0].Attributes != null && nodes[0].Attributes["NotOnOrAfter"] != null)
+            {
+                value = nodes[0].Attributes["NotOnOrAfter"].Value;
+            }
+            return value != null ? DateTime.Parse(value) : (DateTime?)null;
+        }
+
+        public string GetNameID()
+        {
+            XmlNamespaceManager manager = new XmlNamespaceManager(xmlDoc.NameTable);
+            manager.AddNamespace("ds", SignedXml.XmlDsigNamespaceUrl);
+            manager.AddNamespace("saml", "urn:oasis:names:tc:SAML:2.0:assertion");
+            manager.AddNamespace("samlp", "urn:oasis:names:tc:SAML:2.0:protocol");
+
+            XmlNode node = xmlDoc.SelectSingleNode("/samlp:Response/saml:Assertion/saml:Subject/saml:NameID", manager);
+            return node.InnerText;
+        }
+    }
+}
