+++ SpidNetSdk/SpidNetSdk/Saml2/SPIDSaml.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Web;
 
 namespace SpidNetSdk.Saml2
